refactor: split lab offload helpers into modules
