+++ src/core/runner/lab_args.rs
@@ -1,0 +1,76 @@
+use std::path::PathBuf;
+
+use crate::core::{Error, Result};
+
+pub(super) fn lab_offload_source_path(args: &[String]) -> Result<PathBuf> {
+    let mut iter = args.iter().skip(1).peekable();
+    while let Some(arg) = iter.next() {
+        if arg == "--" {
+            break;
+        }
+        if arg == "--path" {
+            let value = iter.next().ok_or_else(|| {
+                Error::validation_invalid_argument(
+                    "path",
+                    "--path requires a value before Lab offload can sync the workspace",
+                    None,
+                    None,
+                )
+            })?;
+            return Ok(PathBuf::from(shellexpand::tilde(value).to_string()));
+        }
+        if let Some(value) = arg.strip_prefix("--path=") {
+            return Ok(PathBuf::from(shellexpand::tilde(value).to_string()));
+        }
+    }
+
+    std::env::current_dir()
+        .map_err(|err| Error::internal_io(err.to_string(), Some("read cwd".to_string())))
+}
+
+pub(super) fn rewrite_lab_offload_args(args: &[String], remote_path: &str) -> Vec<String> {
+    let mut stripped = Vec::with_capacity(args.len());
+    let mut iter = args.iter().peekable();
+    let mut passthrough = false;
+    let has_force_hot = args.iter().any(|arg| arg == "--force-hot");
+    while let Some(arg) = iter.next() {
+        if passthrough {
+            stripped.push(arg.clone());
+            continue;
+        }
+        if arg == "--" {
+            passthrough = true;
+            stripped.push(arg.clone());
+            continue;
+        }
+        if arg == "--path" {
+            stripped.push(arg.clone());
+            let _ = iter.next();
+            stripped.push(remote_path.to_string());
+            continue;
+        }
+        if arg.starts_with("--path=") {
+            stripped.push(format!("--path={remote_path}"));
+            continue;
+        }
+        if arg == "--runner" {
+            let _ = iter.next();
+            continue;
+        }
+        if arg.starts_with("--runner=") {
+            continue;
+        }
+        if arg == "--output" {
+            let _ = iter.next();
+            continue;
+        }
+        if arg.starts_with("--output=") {
+            continue;
+        }
+        stripped.push(arg.clone());
+    }
+    if !has_force_hot {
+        stripped.insert(1, "--force-hot".to_string());
+    }
+    stripped
+}
