+++ src/core/runner/mod.rs
@@ -19,8 +19,12 @@ mod evidence;
 mod execution;
 mod lab;
 mod lab_apply;
+mod lab_args;
+mod lab_capabilities;
 mod lab_command;
 mod lab_env;
+mod lab_plan;
+mod lab_selection;
 mod lab_workspaces;
 mod offload_changed_since;
 mod offload_metadata;
